+++ apps/main/src/features/auth/utils/models/input/confirmation.body.input-model.ts
@@ -1,11 +1,13 @@
-import { Transform, TransformFnParams } from "class-transformer"
-import { IsNotEmpty, IsString, IsUUID } from "class-validator"
+import { Transform } from 'class-transformer'
+import { IsNotEmpty, IsString, IsUUID } from 'class-validator'
 import { trimTransformer } from '../../../../../../../infrastructure/utils/trim-transformer'
+import { ApiProperty } from '@nestjs/swagger'
 
 export class ConfirmationBodyInputModel {
-  @Transform(({ value }) => trimTransformer(value, "code"))
-  @IsString()
-  @IsNotEmpty()
-  @IsUUID()
-  code: string
+	@ApiProperty()
+	@Transform(({ value }) => trimTransformer(value, 'code'))
+	@IsString()
+	@IsNotEmpty()
+	@IsUUID()
+	code: string
 }
