+++ apps/main/src/features/auth/utils/models/input/email-confirmation-resend.body.input-model.ts
@@ -1,12 +1,14 @@
-import { IsEmail, IsString, Matches } from "class-validator"
-import { EMAIL_REGISTRATION_REGEX } from "../../../../../../../infrastructure/utils/constants"
-import { Transform, TransformFnParams } from "class-transformer"
+import { IsEmail, IsString, Matches } from 'class-validator'
+import { EMAIL_REGISTRATION_REGEX } from '../../../../../../../infrastructure/utils/constants'
+import { Transform } from 'class-transformer'
 import { trimTransformer } from '../../../../../../../infrastructure/utils/trim-transformer'
+import { ApiProperty } from '@nestjs/swagger'
 
 export class EmailConfirmationResendBodyInputModel {
-  @Transform(({ value }) => trimTransformer(value, "email"))
-  @IsString()
-  @IsEmail()
-  @Matches(EMAIL_REGISTRATION_REGEX)
-  email: string
+	@ApiProperty()
+	@Transform(({ value }) => trimTransformer(value, 'email'))
+	@IsString()
+	@IsEmail()
+	@Matches(EMAIL_REGISTRATION_REGEX)
+	email: string
 }
