+++ apps/main/src/features/auth/utils/models/input/new-password.body.input-model.ts
@@ -1,16 +1,25 @@
-import { IsNotEmpty, IsString, Length } from "class-validator"
-import { Transform, TransformFnParams } from "class-transformer"
+import { IsNotEmpty, IsString, Length } from 'class-validator'
+import { Transform } from 'class-transformer'
 import { trimTransformer } from '../../../../../../../infrastructure/utils/trim-transformer'
-import { PASSWORD_MAX_LENGTH, PASSWORD_MIN_LENGTH } from '../../../../../../../infrastructure/utils/constants'
+import {
+	PASSWORD_MAX_LENGTH,
+	PASSWORD_MIN_LENGTH
+} from '../../../../../../../infrastructure/utils/constants'
+import { ApiProperty } from '@nestjs/swagger'
 
 export class NewPasswordBodyInputModel {
-  @Transform(({ value }) => trimTransformer(value, "newPassword"))
-  @IsString()
-  @Length(PASSWORD_MIN_LENGTH, PASSWORD_MAX_LENGTH)
-  newPassword: string
+	@ApiProperty({
+		minLength: PASSWORD_MIN_LENGTH,
+		maxLength: PASSWORD_MAX_LENGTH
+	})
+	@Transform(({ value }) => trimTransformer(value, 'newPassword'))
+	@IsString()
+	@Length(PASSWORD_MIN_LENGTH, PASSWORD_MAX_LENGTH)
+	newPassword: string
 
-  @Transform(({ value }) => trimTransformer(value, "recoveryCode"))
-  @IsString()
-  @IsNotEmpty()
-  recoveryCode: string
+	@ApiProperty()
+	@Transform(({ value }) => trimTransformer(value, 'recoveryCode'))
+	@IsString()
+	@IsNotEmpty()
+	recoveryCode: string
 }
